Refactor assetTools, TODO: refactor HOCS/hooks
