+++ src/selectors/assets.js
@@ -1,5 +1,6 @@
 import path from 'path';
-import { get, keys, uniq, flatMap } from 'lodash';
+import { get } from 'lodash';
+import { getNetworkVariables } from '@app/other/protocols/assetTools';
 import { getActiveProtocolMeta } from './protocols';
 import { getAssetManifest } from './protocol';
 
@@ -20,60 +21,20 @@ export const getAssetPath = (state, dataSource) => {
   return assetPath;
 };
 
-// const getNeworkFromList = (nodes = [], edges = []) => {
-//   const nodesWithAttributes = nodes.map(
-//     node => ({ attributes: { ...node } }),
-//   );
-//   const edgesWithAttributes = edges.map(
-//     edge => ({ attributes: { ...edge } }),
-//   );
-
-//   return {
-//     nodes: nodesWithAttributes,
-//     edges: edgesWithAttributes,
-//   };
-// };
-
-// /**
-//  * Read external asset data and return as json network object
-//  *
-//  * @param {string} assetPath path to file on disk
-//  */
-// export const readExternalData = async (assetPath) => {
-//   const extname = path.extname(assetPath).toLowerCase();
-
-//   switch (extname) {
-//     case '.csv': {
-//       const csvData = await fs.readFile(assetPath);
-//       const items = await csv().fromString(csvData.toString('utf8'));
-//       // mock up network object that matches expected json format
-//       const network = getNeworkFromList(items);
-//       return network;
-//     }
-//     case '.json':
-//       return fs.readJson(assetPath);
-//     default:
-//       throw Error(`Unrecognized format '${extname}'`);
-//   }
-// };
-
 /**
  * Extract all unique variables from an external data network asset
  *
- * @param {Object} externalData A network object { nodes: [], edges: [] }
+ * @param {Object} dataSource id of entry in assetManifest
  */
-export const getVariablesFromAsset = (state, dataSource) => {
+export const getNetworkAssetVariables = (state, dataSource, asOptions = true) => {
   const assetPath = getAssetPath(state, dataSource);
-  //   const variableOptions = uniqueAttributes
-  // .map(attribute => ({ label: attribute, value: attribute }));
-  const variables = getVariableNamesFromAsset(assetPath);
-  // const items = externalData[entity] || [];
-  // const allAttributes = flatMap(
-  //   items,
-  //   item => (item.attributes && keys(item.attributes)) || [],
-  // );
-  // const uniqueAttributes = uniq(allAttributes);
+  const variables = getNetworkVariables(assetPath);
 
+  if (asOptions) {
+    const variableOptions = variables
+      .map(attribute => ({ label: attribute, value: attribute }));
+    return variableOptions;
+  }
 
-  // return variableOptions;
+  return variables;
 };
